# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/singlePerson/addSinglePersonAndMatch": {
            "post": {
                "description": "新增匹配",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "singlePerson"
                ],
                "summary": "新增匹配",
                "parameters": [
                    {
                        "description": "新增匹配",
                        "name": "request.AddSinglePersonAndMatchReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddSinglePersonAndMatchReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "新增成功",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/singlePerson/querySinglePeople": {
            "get": {
                "description": "查詢匹配",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "singlePerson"
                ],
                "summary": "查詢匹配",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "可匹配數量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "1": {
                        "description": "查詢可匹配數量",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.SinglePerson"
                            }
                        }
                    }
                }
            }
        },
        "/singlePerson/removeSinglePerson": {
            "delete": {
                "description": "移除匹配",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "singlePerson"
                ],
                "summary": "移除匹配",
                "parameters": [
                    {
                        "description": "userId",
                        "name": "request.RemoveSinglePersonReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RemoveSinglePersonReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "移除成功",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.SinglePerson": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "性別 0:男 1:女",
                    "type": "string"
                },
                "height": {
                    "description": "身高",
                    "type": "number"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "number_want_date": {
                    "description": "人數",
                    "type": "integer"
                },
                "user_id": {
                    "description": "id",
                    "type": "integer"
                }
            }
        },
        "request.AddSinglePersonAndMatchReq": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "性別 0:男 1:女",
                    "type": "string"
                },
                "height": {
                    "description": "身高",
                    "type": "number"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "number_want_date": {
                    "description": "人數",
                    "type": "integer"
                }
            }
        },
        "request.RemoveSinglePersonReq": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "1:成功 2:匹配成功 -1:錯誤 -2:匹配失敗",
                    "type": "string"
                },
                "data": {
                    "description": "匹配資料"
                },
                "err": {
                    "description": "錯誤訊息",
                    "type": "string"
                }
            }
        }
    }
}